feat: use new notifacation schema
