+++ packages/app/hooks/auth/use-wallet.web.ts
@@ -39,13 +39,13 @@ const useWallet = (): UseWalletReturnType => {
   const connected =
     (wagmiData.isConnected && !!wagmiSigner?.provider && !!chain) || isMagic;
 
-  console.log(
-    "wallet connection status ",
-    wagmiData,
-    wagmiSigner,
-    chain,
-    connected
-  );
+  // console.log(
+  //   "wallet connection status ",
+  //   wagmiData,
+  //   wagmiSigner,
+  //   chain,
+  //   connected
+  // );
 
   return {
     address,
