+++ packages/app/lib/utilities.ts
@@ -1,6 +1,6 @@
 import removeMd from "remove-markdown";
 
-import { CONTRACTS } from "./constants";
+import { CHAIN_IDENTIFIERS, CONTRACTS } from "./constants";
 
 export const removeTags = (str) => {
   if (str === null || str === "") return false;
@@ -128,3 +128,9 @@ export const buildFormData = (data) => {
 };
 
 export const toWei = (number: number) => Math.floor(number * 10 ** 9);
+
+export const findTokenChainName = (chainId: string) =>
+  Object.keys(CHAIN_IDENTIFIERS).find(
+    (key: string) =>
+      CHAIN_IDENTIFIERS[key as keyof typeof CHAIN_IDENTIFIERS] == chainId
+  );
